requirements.txt
fastapi
uvicorn
pandas
python-dotenv
langchain
langchain-openai
langchain-chroma
langchain-experimental
pypdf
tiktoken